Create MusicPlayerTest.py
Music player that plays local files. Allows you to input a search and finds it. READ - only works if you type the exact name of file without the .mp3 extension, only works for .mp3, currently only works on mac because of file directory, will play the whole song but the program will not end until you type STOP, doesn't print message when file not found. Currently working on it to see if it something we can actually use in the chatbot.